fix: remove dead code for index nested start keys (#24)
When using index nesting, the parse loop visits `[` and computes keys
_only if_ we didn't just pass a `]`.

This is to capture the start of such a key. For example, `foo[bar][baz]`
would result in the only occurrence of this being `foo[`.

Since a syntactically valid indexed key should only have one of these
(at the start), `lastKey` should always be `undefined` at that point. So
we can remove the check.