Backport to v1 (#58)
* fix: ignore undefined values in stringify (#51)

* chore: add funding.yml (#53)

* chore: drop rimraf (#55)

* feat: inline decode-uri-component (#56)

This dependency rarely changes and is unlikely to be pulled in by a
sibling dependency. So it probably makes sense to inline it.

To avoid having to setup a bundle for what's basically a static file, I
have simply pulled the module into the repo.

* feat: allow bracket properties in js syntax (#57)

Allows `foo[bar]` in `js` nesting syntax, seeing as it is meant to be
js-like after all.

This means the following is valid:

```
foo[bar]=123
// { foo: { bar: 123 } }

foo.bar[baz]=123
// { foo: { bar: { baz: 123 } } }

foo[0][bar].baz=123
// { foo: [ { bar: { baz: 123 } } ] }
```

---------

Co-authored-by: RikSchefferAmsterdam <104840047+RikSchefferAmsterdam@users.noreply.github.com>